Only handle Cookies in Browser

+++ Frontend/src/app/app.component.ts
@@ -49,28 +49,30 @@ export class AppComponent implements OnInit, OnDestroy {
     this.registerIcon("twitter", './assets/svg/twitter.svg');
     this.registerIcon("apple", './assets/svg/apple.svg');
 
-    const cookieConsentStatus = this.cookieService.get("cookieconsent_status");
-    if (cookieConsentStatus && cookieConsentStatus === 'allow') {
-      this.initializeGoogleAnalytics();
-    } else if (cookieConsentStatus && cookieConsentStatus === 'deny') {
-      window['ga-disable-' + environment.gaTrackingCode] = true;
-    }
-
-    this.statusChangeSubscription = this.cookieConsentService.statusChange$.subscribe(async (result: NgcStatusChangeEvent) => {
-      if (result.status) {
-        if (result.status === 'allow') {
-          this.initializeGoogleAnalytics();
-        } else if (result.status === 'deny') {
-          window['ga-disable-' + environment.gaTrackingCode] = true;
-          for (const key in this.cookieService.getAll()) {
-            if (key !== "cookieconsent_status") {
-              this.cookieService.delete(key);
-            }
-          }
-          window.location.reload();
-        }
+    if (!isPlatformServer(this.platformId)) {
+      const cookieConsentStatus = this.cookieService.get("cookieconsent_status");
+      if (cookieConsentStatus && cookieConsentStatus === 'allow') {
+        this.initializeGoogleAnalytics();
+      } else if (cookieConsentStatus && cookieConsentStatus === 'deny') {
+        window['ga-disable-' + environment.gaTrackingCode] = true;
       }
-    });
+
+      this.statusChangeSubscription = this.cookieConsentService.statusChange$.subscribe(async (result: NgcStatusChangeEvent) => {
+        if (result.status) {
+          if (result.status === 'allow') {
+            this.initializeGoogleAnalytics();
+          } else if (result.status === 'deny') {
+            window['ga-disable-' + environment.gaTrackingCode] = true;
+            for (const key in this.cookieService.getAll()) {
+              if (key !== "cookieconsent_status") {
+                this.cookieService.delete(key);
+              }
+            }
+            window.location.reload();
+          }
+        }
+      });
+    }
   }
 
   ngOnDestroy(): void {
